Made simple Box2D rectangular bodies functional
Defined the Body item, which can only be used as a child of a World
item. The standard x, y, width and height properties can be used to
define its position and size, and it will automatically determine its
mass.

The World item has been extended with properties to set the time step,
the number of iterations for each time step and the time between each
frame.