Added support for defining the fixtures of a body
In Box2D, one body is made up of one or more fixtures, where each
fixture has its own density, friction and restitution. Also, fixtures
can be either a convex polygon or a circle.

This change adds support for defining these fixtures in QML, allowing
multiple boxes to be used to make up a single body as well as adding
support for circular shapes. There is no way to define an arbitrary
convex polygon yet, though.

It is now also possible to specify whether a fixture is a sensor, though
no hooks are provided yet to obtain any collision information.