Fixed assertion failure when creating bodies with Component.createObject
QDeclarativeComponent::createObject causes the parent to be set before a
component is complete. In that case the anchors have not been evaluated
yet, which caused an assertion failure when using anchors with a Box
component, since its polygon was still invalid.

Solved this issue by delaying the creation of the internal body until
the Body component is complete.