Used a QAbstractAnimation to synchronize stepping to frames
Until now a timer was used with an interval of 16 ms. While this
synchronizes with a 60 fps framerate most of the time, it will regularly
not match up correctly resulting in frames being skipped.

Now a QAbstractAnimation is used instead. Its 'updateCurrentTime' method
is called for each frame so it is perfectly synchronized, resulting in
perfect smooth animation (as long as the step can be done fast enough).

The 'frameTime' property is no longer needed and was removed.