Simplified initial registration of bodies
No need to put them in a list first.

I do find the support for bodies that are not direct children of the
world a little problematic. Apart from coordinate system issues,
currently also the world only notices the addition and removal of bodies
that are its own children. When a body is deeper down in the hierarchy,
it is registered on initialization but it will fail to do so when added
dynamically.