Fixed crash when starting the 'mouse' example
Since the world is not known during initialization but only once things
are being created, the joints also do not know the value of
pixelsPerMeter during initialization.

The solution implemented here delays the creation of a b2JointDef
instance (along with any necessary conversions) until the point where
the joint is actually being created.

Only the MouseJoint has been fixed, the other joints would have to be
adapted in the same way.