Introduced Body.target
This allows using any existing item as the target for physics emulation,
rather than requiring such items to be wrapped by a Body item.

Since the World object still needs to find the Body instances, a dynamic
property is set on items set as a Body.target which is used instead of a
qobject_cast.