Adjusted all examples to the QObject-based API
The main thing here is that the World, Body and fixture components are
no longer derived from Item but from QObject. This means they don't have
a parent item and they don't have anchors.

Many examples now use a shared PhysicsItem component that is similar to
the previous Body component, in that it combines an item with a Body
instance that has the item as its target.