Allow using a different name for the static import
QML currently doesn't allow doing qmlRegisterType on a plugin URI
already defined in an installed plugin. It makes sense to use a
different URI for the box2d import when linking it statically to
avoid a fatal QML load error on systems with this plugin installed.

Example:
    Box2DPlugin plugin;
    plugin.registerTypes("Box2DStatic");

This reuses the STATIC_PLUGIN_BOX2D define in box2dplugin.h and
define it properly in box2d-static.pri. The assert is only checked
on non-static builds.