Include some moc-generated files
Apart from reducing compile time slightly, this resolves an issue in
Qt 6 where a compile-time check in the generated files fails when
pointers are used in Q_PROPERTY values of classes that are only
forward-declared.

This worked for the Qbs build, but the include was wrong for the qmake
build due to the usage of "MOC_DIR = .moc". I've just removed that line
since I could not find a way to do that in Qbs, and usually builds are
done out-of-source anyway.